[Maintenance] [th04] Get rid of th04.hpp
Oh wow, we never used that one; I only added it in 042b780 to make it
easier for tool authors to #include "th04/resident.hpp". Bad idea –
if we don't ensure that it *actually works* (and it didn't, as of
d1f3dcd), it's of no help to them either.

Part of P0147, funded by -Tom- and Ember2528.