[Decompilation] [th01] Konngara: Pattern 7/12
The one where Konngara shoots rays in a semicircle motion from their
sword to the edge of the playfield, with pellets raining down from the
ray's end point.

Insert "if… else if… else if…" joke. And hey, I've finally had the idea
to use function-local enums in place of #define!

Part of P0154, funded by Ember2528.