[Decompilation] [th01] Sariel: Spawn ray rendering
So if Sariel's version of this function is worse than Konngara's, can
we conclude that Sariel's code was written first?

(Edit, after having decompiled more of the fight: Sariel's functions
are predominantly `near` in contrast to Konngara's 100% `far` ones,
which is just as much of an indicator for the opposite.)

Part of P0174, funded by Ember2528.