[Decompilation] [th01] Sariel: Lowered / initial wand sprite blitting
Surprise, this .PTN sprite isn't actually just the background, but a
complete opaque copy of the 128×128 pixels in the .GRP image that
contain the initial position of the wand. This is what allows Sariel
to lower it again after blitting the raised sprites from BOSS6_2.BOS.

Part of P0175, funded by Ember2528.