[Maintenance] [th03/th04/th05] Drop the `_lebcd` suffix from score variables
What is this, reversed Hungarian notation? For gaiji strings, it makes
sense because there's no difference in the access code to regular
strings, but you can't really do anything with a union itself.
(Besides passing it to a function in a type-safe way, but that's cool.)

Part of P0186, funded by [Anonymous] and Blue Bolt.