[Decompilation] [th04/th05] Dialog: Script freeing function
One of the rare cases where explicitly spelling out the FP_SEG() cast
is better than just calling the C++ wrapper: To drive home the point
that this code relies on `far` pointer semantics. Running the dialog
script performs arithmetic on only the offset part of this pointer, and
the segment part must remain unchanged for this hmem_free() call to
work as intended.

Completes P0186, funded by [Anonymous] and Blue Bolt.