[Decompilation] [th05] Shinki: Pattern 1/11
The one where Shinki fires bullets in a ring, with each new ring fired
at an increasing or decreasing angle, for a curved effect. The first
pattern you see in the fight.

At least the pattern functions are short and sweet for the most part.

Part of P0190, funded by nrook.