[Naming] [th04/th05] Boss-specific names for backdrop colorfill functions
Once we decompile whatever can be decompiled of those, we're going to
use the same coordinate constants for both these and the backdrop
image. Ideally, we can even macro (or inline-function) away the entire
implementation, so that the redundancy between TH04's and TH05's Stage
4 bosses won't matter as much.

Part of P0189, funded by Arandui and Lmocinemod.