[Maintenance] [th05] Custom entities: Introduce separate template structures
Since the game uses global state to define its patterns, we'd really
like to see immediately which of the structure fields are relevant for
spawning bullets. This makes it easier to spot which of the patterns
rely on fields set in previous patterns – and there are several that
do.

Part of P0190, funded by nrook.