[Reverse-engineering] [th04/th05] Bosses: Patterns seen / defeat bonus flag
There's the meaning I was looking for in 6ff427a! Far from everything
that's done with this "mode change" variable though – and as we've seen
in 426a531, it's mostly a waste of time to immediately RE all meanings
of this variable across all bosses. So, let's keep using the raw
"phase_state" variable in ASM land.

Part of P0191, funded by nrook.