[Naming] [th01] Elis: Generic pattern state variable
Yup, assigning this one to Elis, even though its proximity to the rest
of the generic boss variables might suggest that it was intended as a
generic boss state variable. Nope, that idea has already demonstrated
its uglyness in the previous TH04 and TH05 pushes…

Part of P0193, funded by Ember2528.