[Decompilation] [th01] Elis: Pattern 7/13
The one where Elis' bat form fires a slow spread group every 16 frames.
Interestingly, it's easiest on Lunatic, where the pellets are always
fired downwards and you just have to not stand below Elis. On every
other difficulty, the pellets are aimed to the player.

And what's this, an appropriately used local variable?!

Part of P0195, funded by Yanga.