[Naming] [th01] Stage objects: Rename the background+object blitting function
All throughout this codebase, `X_put_8()` implies that just the single
sprite of X is blitted on top of whatever is in VRAM at the given
position. This function blits more than one sprite though. (Also,
portal rendering is going to introduce a helper function that does only
blit a single thing.)

Part of P0203, funded by [Anonymous] and GhostRiderCog.