[Reverse-engineering] [th01] YuugenMagan: Eye distances
Turns out that we can define the magic numbers used in the pentagram
shrink effect in terms of the distances between the eyes and the
pentagram's target position in the center of the playfield. Therefore,
specifying the eye positions in raw pixel coordinates won't give the
full picture (who would have thought).

Part of P0208, funded by GhostPhanom and Yanga.