[Decompilation] [th01] FUUIN.EXE: main()
The final secret of this executable: With a loaded MDRV2, enter

	fuuin t1

on the DOS prompt to view the 魔界/Makai Good Ending, or

	fuuin t

for the 地獄/Jigoku Good Ending.

(Bad Endings can only be accessed by pre-allocating the TH01 resident
structure and setting one of the `continues_per_scene` array members to
a nonzero value.)

Part of P0213, funded by Ember2528 and GhostRiderCog.