[Debloating] [th01] Shootout lasers: Remove the default constructor
It might be good form, but since these lasers are statically allocated,
the previous constructor pulled in _vector_new_() together with a bunch
of exception handling code that is *much* harder to get rid of than the
usual bloat. All members are automatically zeroed anyway, by virtue of
*being* static.

Part of P0229, funded by Ember2528.