[Debloating] [th01] Pellets: Include delay clouds in the counter after reset
Since this counter was originally only used as a shortcut for skipping
the spawn loop once we reached the pellet cap, we can simply "repair"
it to both preserve the quirk *and* allow it to be used to skip other
loops, including the rendering one.
Over on the `anniversary` branch, this fixes a bug introduced in the
P0234 release that caused delay cloud pellets to not be unblitted after
a reset, leading to them drawing white streaks across their trajectory.