[Naming] Rename vector1_at() to polar()
"1D polar vectors" are a weird concept to wrap your head around,
especially if their constructor function takes a sine or cosine ratio
instead of an angle. This function is way closer to the polar_*()
family, which has these convenient `_x` and _y` suffixes that we could
just remove for this generic function to fit into there.

Part of P0244, funded by Blue Bolt and [Anonymous].