[Research] Discover the reason for incorrect clipping of large sprites
It's going to show up again for TH04's Stage 4 midboss, as well as the
big explosion sprite, and this turns out to be the reason why ZUN did
it consistently. master.lib being lazy is enough of a reason to upgrade
each of these instances to a fixable bug.

Part of P0245, funded by [Anonymous], Blue Bolt, Ember2528, and Yanga.