[Decompilation] [th03/th04/th05] Music Room: B plane snapping and blitting
Might as well give a more descriptive name to the variable than
something involving the generic `screen` (which we've started using for
coordinates) and the inconsistent `back` (which always means "back page
of VRAM", which has nothing to do with this bitplane).
Also, turns out that this was in fact the only place that did a 32-wide
memcpy() of a full VRAM bitplane.

Part of P0264, funded by [Anonymous] and Blue Bolt.