[Reverse-engineering] [th02] Total and stage-specific miss and bomb counters
The stage bonus functions require the stage-specific ones, which are
updated next to the total ones, which are required by the skill
calculation that we need to explain what we just saw in the verdict
screen… What a coincidence! (Yes, this actually wasn't planned.)

Part of P0279, funded by Yanga and Blue Bolt.