[Maintenance] Spell out master.lib's RNG functions
`irand()` and `rand()` are only identical on Borland compilers. The
specific RNG implementation is an integral part of the original game
logic, so we actually never want to use the generic libc `rand()`. Any
future replays recorded on PC-98 builds should play back identically on
whatever system we port the games to, after all.

Part of P0284, funded by [Anonymous] and Blue Bolt.