[Build] Rules: Batch TCC invocations with identical command lines
Compiling as many source files as possible within a single TCC process
is the single most important build time optimization we can do. Sadly,
it won't work with Tup's parallel nature, but it can really shine in a
dumb batch build. This is the reason why Windows 9x can compile this
entire codebase on a single core in half the time it takes MS-DOS
Player to do the same on 64-bit Windows, on the same hardware.

That weird double-spacing, however, is the most ridiculous workaround
for a bug I have seen in a long while…

Completes P0282, funded by [Anonymous].