[Decompilation] [th04/th05] High Score screen: GENSOU.SCR loading
For all intents and purposes, this code belongs to the rest of the
`GENSOU.SCR` implementation code in `th04/formats/scoredat/`. That
eyesore of a `rank` variable is the only thing that maybe sort of
relates these functions to the High Score screen. We could even rename
that variable to something like `hiscore_rank` to make the distinction
more obvious if it weren't for the fact that the `MAIN.EXE` version
does in fact use the global [rank]. 🤦
In the end though, it's code generation reasons that force this code
into the translation unit of the respective binary's High Score screen,
because the `GENSOU.SCR\0` string is 11 bytes long and not followed by
a padding alignment byte.

Part of P0298, funded by [Anonymous], Blue Bolt, and Splashman.