[Maintenance] [th03] Limit the previous `player.hpp` to the stuff structure
Otherwise, it would declare more and more subsystems that store
per-player data but are implemented in wildly different places.
It's very unfortunate that ZUN didn't go for either full-on decoupled
SoA or full-on giant structure of everything, and instead decided to
use both approaches at the same time. These "stuff" structures resist
any attempt at neatly organizing the reconstructed code, and all we can
do is to somehow explain it with comments.

Part of P0312, funded by [Anonymous], Blue Bolt, and iruleatgames.