[Decompilation] [th03] Staff Roll / Verdict: Snowflake sprite blitting
Too bad that "snowflake" does not fit into 8.3 characters, so "flake"
it is once again, just as in TH01.
Also, ZUN went for runtime-shifted blitting for once, but still
exchanged 32 bytes of "sprite data" for the comfort of not having to
use `MOVZX` or `XOR AH, AH`?

Part of P0312, funded by [Anonymous], Blue Bolt, and iruleatgames.