[Platform] [PC-98] Blitter: Move low-level code into a separate header
Regular game code should not be bothered with all these declarations.
Also, we're going require a per-game translation unit to instantiate
blitters for the widths a game actually wants to use – generating all
of them would not only bloat the binary (and the planned dedicated
render code segment), but also take ~10 seconds to compile.

Part of P0314, funded by [Anonymous].