[Portability] Graphics: Specify direct .PI blitting to VRAM
Not sure how long we're going to keep this – I'd like to get rid of all
VRAM tricks in the long run, and the cases where games just want to
blit their background once and then never do something with it again
are rare. But this is a nice start for both integrating PiLoad in later
games, and moves us toward more explicit palette handling.

Part of P0316, funded by [Anonymous], Ember2528, and Congrio.