[Platform] [PC-98] Blitter: Optimize for emulators by default
Tests on real hardware confirmed that even the slowest Pentium models
are bottlenecked by the GDC rather than the CPU. Their superscalar
microarchitecture erases *any* difference between blitting methods, so
it doesn't matter what we choose as far as those are concerned.

Benchmark results from 486 models would be interesting, but I haven't
received any so far.

That only leaves the 386 as a worthwhile target to optimize for – and
as we've seen earlier, that's also what the Neko Project family sort of
targets. Hence, this is also the best choice for the majority of
players.
And as a nice bonus, string unit instructions are decently small at 1
byte per 16 pixels.

Part of P0314, funded by [Anonymous].