[Platform] [PC-98] Blitter: Remove the per-width `or` function pointer
We're later going to reveal that we *really* want to use the EGC for
this job, and we can never realistically run these games on systems
without an EGC anyway. The check/shift benchmark is the only program
that needs this operation to show off how slow it is, and that one
swaps out the function pointers anyway, so it can just re-use `write`.

Part of P0314, funded by [Anonymous].