[Platform] [PC-98] Blitter: Add marching unit writing to the offscreen area
We can blast horizontally uncropped sprites to the offscreen area of
VRAM with a simple `REP MOVS`, as this is just memory we want to use as
efficiently as possible and don't care about a sprite's spatial layout.
For cropped sprites, however, there's no way around pointer addition
after every line. That's why this commit uses the REP-less MOVS
instructions as they perform better with the low widths we're likely to
crop our sprites to.

Part of P0314, funded by [Anonymous].