[Platform] [PC-98] EGC: Add a RAII base class for direct control
Neko Project 21/W is the only platform with predicable execution
timings we can target, and it emulates a 386 where even near calls are
very costly. So we're once again going for the inlined method style
we use for the GRCG. Symmetry!

Part of P0316, funded by [Anonymous], Ember2528, and Congrio.