[Decompilation] [th04] Title screen animation
The stylish effect of separately whiting in the black outlines is
definitely not appreciated enough. The whole code sure does silly
things with the .PI palette though, which is why it made sense to
quickly decompile this one before we start debloating all these
binaries.
Also, that overloading of variables…

Part of P0318, funded by [Anonymous] and Ember2528.