[Debloating] Compile position-independent binaries with semi-optimal flags
Pascal calls are too disruptive for the remaining not yet decompiled
code, but everything else retains the ABI and removes a few bytes from
the binary.
Skipping TH02's and TH03's `OP.EXE` for now though, as the flags would
otherwise be applied to code that is shared with `MAIN.EXE` and mustn't
be modified right now.

Part of P0313, funded by Ember2528 and Splashman.