[Debloating] [th02] Main menu: Retain the background image
Boom! Well-configured emulators now run the switch from the High Score
viewer back to the main menu now in exactly the 20 frames that ZUN
specified.

Just like in TH04 and TH05, we also unfortunately have to raise the
memory limit here. In this game, however, we need a whopping and very
specific additional 80,416 bytes of memory to incorporate
• the 8,192-byte gaiji backup,
• the three title animation cels (3 × 54,720 = 164,160 bytes),
• the 16,384-byte .PI load buffer (which, as it's allocated *before* we
  allocate any .PI, will necessarily fragment away that memory),
• the 128,000 bytes of the background image, and
• the 2,560 bytes loaded from `op_h.bft`. (Which would have been
  11,776 bytes if we didn't free the `super_buffer` in an earlier
  commit.)

Moving the `op2.pi` load call after the end of the flash animation
affects the original frame timing on real PC-98 systems since there is
no intentional delay that could hide the loading.

But even this increase is fine for TH02. The game's original `MAIN.EXE`
takes up 164,864 bytes of static memory and calls for 288,000 bytes of
heap, bringing the game's total memory requirement to 452,864 bytes
excluding PMD. Thus, asking for 336,416 bytes of heap is no problem at
all as long as `OP.EXE`'s static memory consumption is ≤116,448 bytes.
And even after we'll have merged `MAINE.EXE`, we'll stay way below 70
KB, even with the low-effort merged binary that this delivery will
ship.

Part of P0320, funded by [Anonymous] and Ember2528.