[Portability] Graphics: Specify 1bpp surfaces
Remains the cleanest way of expressing the Music Room polygon effect,
and will also be very useful for future pre-rendered text. Note how the
plane parameter for the `write()` method points out that this call is
not meant to be used with the GRCG.

Part of P0316, funded by [Anonymous], Ember2528, and Congrio.