[Maintenance] Remove indentation from preprocessor branches
I decided on this convention for the Shuusou Gyoku codebase earlier
this year and quickly got used to it, to the point that I now find it
hard to read indented preprocessor branches. The previous visual
structure of the code has put me off debloating the Music Room just
because it made the run-time branching seem worse than it actually is.
And with LSPs highlighting the code of the game you're currently
investigating and dimming out the code that game doesn't use, it looks
increasingly bad to suddenly see code jump by one level of indentation
due to a branch that is resolved at compile time.

Part of P0313, funded by Ember2528 and Splashman.