[Landmine] [th04] High Score viewer: Don't glitch on GENSOU.SCR corruption
An unusually large diff for a landmine fix, but it's the mere existence
of `hi2` that causes the issue in the first place. Refactoring the code
*without* simultaneously fixing the issue would have required us to
move away from the global `hi` variable before.

With this fix, Reimu's scores also happen to be properly shown if only
Marisa's are corrupted. Ultimately, this doesn't matter because it's a
landmine after all, but it's still a change in behavior.

Part of P0321, funded by [Anonymous] and Ember2528.