[Debloating] [th02/th03/th04/th05] Group SHARED code at the end
For some reason, interleaving `SHARED` code with game code the way we
must do it on `master` runs the risk of the whole `SHARED` segment
getting put into the first code group, depending on which segments come
before and after. This destroys any CS-relative references and results
in crashes. Grouping all `SHARED` code like this avoids this problem.

Going even further and alphabetically sorting the list might look like
changes for the sake of changes at first. Ultimately though, all of
this code should be compiled into a static library and automatically
pulled into the binaries as needed anyway, removing the need for this
list on all branches.

Part of P0313, funded by Ember2528 and Splashman.