[Debloating] [th02/th03/th04/th05] Music Room: Use `near` pointers for strings
Every `near` pointer not only saves the obvious 2 bytes in the data
segment, but also another 4 for the no longer required relocation in
the MZ header. TH05 is, obviously, the biggest beneficiary here: Its
226 Music Room strings translate to (6 × 226) = 1,356 bytes of bloat
shaved off OP.EXE with this trick alone, and probably a few more due to
the relocation table's 512-byte alignment.

Part of P0322, funded by [Anonymous].