[Debloating] [th02/th03/th04/th05] Music Room: Trim trailing whitespace
Saves another 750 bytes in TH05, in exchange for preventing hex-edited
translations and making even source-level translations into other
ASCII languages slightly awkward… 🤔 But come on, you should be
committing and pushing your translations somewhere.
And really, we'd like to have a less hardcoded and more automatic and
correct kind of centering in the end.

Part of P0322, funded by [Anonymous].