[Platform] [PC-98] VBLANK: Add a mechanism for running code in VBLANK
Of course, this doesn't defend against the hardware leaving VBLANK
while the code is still running. But it's a decent workaround for not
delaying certain GDC mutations by a whole frame on faster systems where
we get more cycles in VBLANK.

Part of P0316, funded by [Anonymous], Ember2528, and Congrio.