[Landmine] [th01/th02/th03] graph_putsa_fx(): Merge implementations
Including `text_extent()` since we'll need it for right alignment on
the `anniversary` branch. Also keeping `FX_CLEAR_BG` and `FX_REVERSE`
because TH01 still needs them. It's all still far from optimized and
portable, so we're definitely going this to revisit this in the future.
And yes, this also fixes the landmine that turns illegal Shift-JIS
codepoints into `=`!

Part of P0323, funded by Ember2528.