[Bug] [th01] GAME.BAT replica: Correctly shut down self-spawned MDRV2
Previously, we would only shut down a self-spawned MDRV2 if the player
quit to DOS from the same process that spawned it, since we kept the
spawn state in a variable local to `main_setup()`. But this would only
be the case when immediately quitting from the Main menu, without ever
playing a gameplay round. The fix (and proper implementation for the
other 4 games) requires the new shutdown mechanism, as well as
repurposing an unused byte inside the resident structure for that spawn
state.

Part of P0322, funded by [Anonymous].