[Debloating] [th02/th03/th04/th05] Remove `_setupio()` from PI binaries
Freeing up not nearly as much as the exception handler freed up
previously, but still a nice 18,260 bytes across the 10 binaries it
applies to. Can't remove it from TH01 for the reason mentioned in the
`novbuf.cpp` comment, but it would have only removed 166 bytes anyway.

Part of P0322, funded by [Anonymous].