[Debloating] [th02/th03/th04/th05] Replicate GAME.BAT
Second iteration of my second attempt at solving this problem.
Test spawns are the only reliable way of handling the various different
PMD versions that ship with the game and that modders might want to
use, so why not do the same for every single TSR we have to spawn,
using a data-driven algorithm. You do *not* want to see the first
attempt.

Part of P0322, funded by [Anonymous].